extract tacticus API to lib
