+++ src/fsd/5-shared/ui/index.ts
@@ -1,3 +1,4 @@
+export * from './contexts';
 export { MenuItemTP } from './menu-item';
 export { Conditional } from './conditional';
 export { usePopUpControls } from './pop-up-controls';
