+++ src/models/constants.ts
@@ -3,9 +3,7 @@
 import { Rank, Rarity, RarityStars, RarityMapper } from '@/fsd/5-shared/model';
 
 import { ICampaignsProgress, Campaign } from '@/fsd/4-entities/campaign';
-import { CharacterReleaseRarity } from '@/fsd/4-entities/character';
-import { CharactersFilterBy } from '@/fsd/4-entities/character/characters-filter-by.enum';
-import { CharactersOrderBy } from '@/fsd/4-entities/character/characters-order-by.enum';
+import { CharactersFilterBy, CharactersOrderBy } from '@/fsd/4-entities/character';
 
 import { GuildWarTeamType, IGWLayoutZone } from 'src/v2/features/guild-war/guild-war.models';
