+++ src/v2/features/tacticus-integration/guild-overview.tsx
@@ -2,12 +2,12 @@ import { AllCommunityModule, ColDef, themeBalham } from 'ag-grid-community';
 import { AgGridReact } from 'ag-grid-react';
 import React, { useState, useEffect } from 'react';
 
-import { IGuildMember } from '@/models/interfaces';
-
-import { getTacticusGuildData } from '@/v2/features/tacticus-integration/tacticus-integration.endpoints';
-import { mapUserIdToName } from '@/v2/features/tacticus-integration/user-id-mapper';
-
-import { TacticusGuild, TacticusGuildMember, TacticusGuildRole } from './tacticus-integration.models';
+import {
+    TacticusGuildRole,
+    TacticusGuild,
+    getTacticusGuildData,
+    TacticusGuildMember,
+} from '@/fsd/5-shared/lib/tacticus-api';
 
 // Helper function to convert role enum to readable string
 const getRoleLabel = (role: TacticusGuildRole): string => {
