+++ src/v2/features/tacticus-integration/tacticus-integration.dialog.tsx
@@ -3,13 +3,13 @@ import React, { useState } from 'react';
 
 import { DialogProps } from 'src/v2/models/dialog.props';
 
+import { updateTacticusApiKey } from '@/fsd/5-shared/lib/tacticus-api';
 import { useAuth } from '@/fsd/5-shared/model';
 import { Button } from '@/fsd/5-shared/ui/button';
 import { useLoader } from '@/fsd/5-shared/ui/contexts';
 import { Checkbox, CheckboxGroup, TextField } from '@/fsd/5-shared/ui/input';
 import { Modal } from '@/fsd/5-shared/ui/modal';
 
-import { updateTacticusApiKey } from './tacticus-integration.endpoints';
 import { useSyncWithTacticus } from './useSyncWithTacticus';
 
 interface Props extends DialogProps {
